Chunk transfer fix
This fixes chunk transfer encoding discovery to adapt to newer
versions of the API. It also introduces a tracing level for debug, which
is used to minimise verbosity at the old debug level. Finally this
also fixes proper debug module output.

This is larger than expected as it has gone through cosmetical reindent.